refactor code to use es6 arrow functions
