+++ modules/RenderBooks.js
@@ -1,5 +1,14 @@
 const article = document.getElementById('bookList');
 const savedBooks = JSON.parse(localStorage.getItem('savedBooks')) || [];
+
+const removeBook = (index) => {
+  savedBooks.splice(index, 1);
+  updateLocalStorage();
+};
+const updateLocalStorage = () => {
+  localStorage.setItem('savedBooks', JSON.stringify(this.savedBooks));
+};
+
 const renderBooks = () => {
   article.innerHTML = '';
   savedBooks.forEach((book, index) => {
